refactor(server): decompose inline tool registration into src/tools/
src/server.ts was registering 12 freelance_* tools inline via
server.tool() calls inside createServer(). At 475 lines it was past
the readability break point — the review flagged it as the single
biggest legibility win available, and the upstream reference
(@modelcontextprotocol/server-everything) uses one-tool-per-file
for its 18 tools.

This commit mirrors the existing src/memory/tools.ts pattern but
takes it one step further: each freelance_* tool lives in its own
~30-line file under src/tools/. An orchestrator in src/tools/index.ts
composes them via a typed deps bundle and exposes a single
registerFreelanceTools(server, deps) entry point that server.ts
calls.

File layout:

  src/tools/
    index.ts            # FreelanceToolDeps + registerFreelanceTools
    list.ts             # freelance_list
    start.ts            # freelance_start
    advance.ts          # freelance_advance
    context-set.ts      # freelance_context_set
    inspect.ts          # freelance_inspect
    reset.ts            # freelance_reset
    guide.ts            # freelance_guide (stateless)
    distill.ts          # freelance_distill (stateless)
    sources-hash.ts     # freelance_sources_hash
    sources-check.ts    # freelance_sources_check
    sources-validate.ts # freelance_sources_validate
    validate.ts         # freelance_validate

Details worth noting:

- `FreelanceToolDeps` exposes the mutable loadErrors via a getter
  (`getLoadErrors: () => ...`), not a snapshot. The watcher in
  server.ts still mutates `currentLoadErrors` on reload, and the
  list tool reads through the getter so it always sees the current
  value without needing to be re-registered.
- Stateless tools (`guide`, `distill`) take just `server`; the rest
  take `(server, deps)` and destructure only what they need.
- `handleError` moved from server.ts to mcp-helpers.ts so it can be
  shared across all tool files. Behavior identical — wraps
  EngineError as a clean message, everything else as "Internal
  error: ...".
- server.ts drops from 475 → 137 lines. It's now:
    imports → ServerOptions type → createServer() (watcher setup,
    server construction, registerFreelanceTools call, memory block,
    return) → startServer().
  The 12 tool definitions are gone from this file entirely.

No behavior change. 576/576 tests pass. Biome check clean. Tool
descriptions are preserved verbatim — the 3-6 line description audit
from the review lands in the next commit.
